refactor: rename file user.ts
